+++ rs-core/random.hpp
@@ -32,8 +32,8 @@ namespace RS {
 
     // LCG functions
 
-    constexpr uint32_t lcg32(uint32_t x) noexcept { return 32310901ul * x + 850757001ul; }
-    constexpr uint64_t lcg64(uint64_t x) noexcept { return 3935559000370003845ull * x + 8831144850135198739ull; }
+    constexpr uint32_t lcg32(uint32_t x) noexcept { return uint32_t(32310901ul * x + 850757001ul); }
+    constexpr uint64_t lcg64(uint64_t x) noexcept { return uint64_t(3935559000370003845ull * x + 8831144850135198739ull); }
 
     class Lcg32:
     public EqualityComparable<Lcg32> {
